timer interrupt STILL not working
Verified that both external interrupts work, so I don't think it's an
issue in where the assembler places the interrupt code. I don't know
where the issue is - either the timer is running and the overflow
interrupts are firing but the ISR is wrong (unlikely; the ISR is written
identically to the working INT0 and INT1 ISRs), the timer is running but
not spitting out overflow interrupts, or the timer isn't running at all
or is resetting before it overflows.